return sheets data merged with people and journalist
