+++ requirements.txt
@@ -13,3 +13,6 @@ gunicorn[gevent]==20.0.4
 openpyxl==3.0.5
 psycopg2==2.8.6
 python-dotenv==0.15.0
+numpy==1.19.3
+pandas==1.1.4
+xlrd==1.2.0
